refact(sidebar-entity-table): moving logic to hook
