+++ dashboard/src/features/entity-table/sidebar-entity-table.tsx
@@ -1,44 +1,29 @@
-import { useMemo } from "react";
+import { Button, DataTableViewOptions, ResizableHandle, ResizablePanel, ResizablePanelGroup, HStack } from "@marzneshin/components";
+import { Server } from "lucide-react";
+import { EntityTableContext, SidebarEntityTableContext } from "./contexts";
 import {
-    Button, DataTableViewOptions,
-    ResizableHandle, ResizablePanel, ResizablePanelGroup,
-    HStack
-} from "@marzneshin/components";
-import { useTranslation } from "react-i18next";
-import { useQuery } from "@tanstack/react-query";
-import {
-    SidebarEntityTableContext,
-    EntityTableContext
-} from "./contexts";
-import {
+    SidebarEntityCardSectionsProps,
     TableFiltering,
     DataTablePagination,
     EntityDataTable,
-    type SidebarEntityCardSectionsProps,
     SidebarEntityTablePopover,
     SidebarEntitySelection
 } from "./components";
 import {
-    useFiltering,
-    usePagination,
-    useEntityTable,
-    useVisibility,
-    useSorting,
-    type SortableSidebarQueryKey,
-    type EntitySidebarQueryKeyType,
-    type FetchEntityReturn,
-    type SidebarQueryKey,
-    type UseRowSelectionReturn,
+    useSidebarEntityTable,
+    EntitySidebarQueryKeyType,
+    FetchEntityReturn,
+    UseRowSelectionReturn,
 } from "./hooks";
-import { useScreenBreakpoint } from "@marzneshin/hooks";
-import { Server } from "lucide-react";
+import { useTranslation } from "react-i18next";
+
 
 interface SidebarEntityTableProps<T, S> {
     fetchEntity: ({ queryKey }: EntitySidebarQueryKeyType) => FetchEntityReturn<T>;
     sidebarEntities: S[];
     setSidebarEntityId: (entity: string | undefined) => void;
     sidebarEntityId?: string;
-    sidebarCardProps: SidebarEntityCardSectionsProps<S>
+    sidebarCardProps: SidebarEntityCardSectionsProps<S>;
     secondaryEntityKey: string;
     columnsFn: any;
     filteredColumn: string;
@@ -51,87 +36,16 @@ interface SidebarEntityTableProps<T, S> {
     onDelete: (entity: T) => void;
 }
 
-export function SidebarEntityTable<T, S>({
-    fetchEntity,
-    columnsFn,
-    filteredColumn,
-    rowSelection,
-    entityKey,
-    manualSorting = false,
-    onCreate,
-    onEdit,
-    onOpen,
-    onDelete,
-    sidebarEntityId,
-    setSidebarEntityId,
-    sidebarEntities,
-    sidebarCardProps,
-    secondaryEntityKey,
-}: SidebarEntityTableProps<T, S>) {
+export function SidebarEntityTable<T, S>(props: SidebarEntityTableProps<T, S>) {
     const { t } = useTranslation();
-    const filtering = useFiltering({ column: filteredColumn });
-    const sorting = useSorting();
-    const visibility = useVisibility();
-    const desktop = useScreenBreakpoint("md");
-    const { onPaginationChange, pageIndex, pageSize } = usePagination();
-
-    const sortedQuery: SortableSidebarQueryKey = [
-        entityKey,
-        sidebarEntityId,
-        secondaryEntityKey,
-        pageIndex,
-        pageSize,
-        filtering.columnFilters,
-        sorting.sorting[0]?.id ? sorting.sorting[0].id : "created_at",
-        sorting.sorting[0]?.desc,
-    ];
-
-    const query: SidebarQueryKey = [
-        entityKey,
-        sidebarEntityId,
-        secondaryEntityKey,
-        pageIndex,
-        pageSize,
-        filtering.columnFilters,
-    ];
-
-    const { data, isPending } = useQuery({
-        queryFn: fetchEntity,
-        queryKey: manualSorting ? sortedQuery : query,
-        initialData: { entity: [], pageCount: 1 },
-    });
-
-    const columns = columnsFn({ onEdit, onDelete, onOpen });
-    const table = useEntityTable({
-        data,
+    const {
+        desktop,
+        table,
+        entityTableContextValue,
+        sidebarEntityTableContextValue,
         columns,
-        pageSize,
-        pageIndex,
-        rowSelection,
-        visibility,
-        sorting,
-        onPaginationChange,
-    });
-
-    const entityTableContextValue = useMemo(
-        () => ({
-            table,
-            data: data.entity,
-            filtering,
-            isLoading: isPending,
-        }),
-        [table, data.entity, filtering, isPending],
-    );
-
-    const sidebarEntityTableContextValue = useMemo(
-        () => ({
-            sidebarEntities,
-            setSidebarEntityId,
-            sidebarEntityId,
-            sidebarCardProps
-        }),
-        [sidebarEntities, setSidebarEntityId, sidebarEntityId, sidebarCardProps],
-    );
+        onCreate
+    } = useSidebarEntityTable(props);
 
     return (
         <EntityTableContext.Provider value={entityTableContextValue}>
@@ -151,7 +65,7 @@ export function SidebarEntityTable<T, S>({
                             )}
                             <DataTableViewOptions table={table} />
                             {onCreate && (
-                                <Button aria-label={`create-${entityKey}`} onClick={onCreate}>
+                                <Button aria-label={`create-${props.entityKey}`} onClick={onCreate}>
                                     {t("create")}
                                 </Button>
                             )}
@@ -165,7 +79,7 @@ export function SidebarEntityTable<T, S>({
                             <ResizableHandle withHandle={desktop} />
                             <ResizablePanel >
                                 <div className="flex flex-col justify-between size-full">
-                                    <EntityDataTable columns={columns} onRowClick={onOpen} />
+                                    <EntityDataTable columns={columns} onRowClick={props.onOpen} />
                                     <DataTablePagination />
                                 </div>
                             </ResizablePanel>
@@ -173,6 +87,6 @@ export function SidebarEntityTable<T, S>({
                     </div>
                 </div>
             </SidebarEntityTableContext.Provider>
-        </EntityTableContext.Provider >
+        </EntityTableContext.Provider>
     );
 }
